Update neovim and make coc-nvim work with snippets
After updating neovim and coc I had to work out a way to have the Tab
key work completion as well as snippet expansion.
The init.vim file now does not use Supertab and UltiSnips anymore. I
changed everything such that coc manages completion as well as snips
which are taken from the UltiSnips files that I used to have. Minor
corrections might have to be made in the future but so far it looks
good.